backend/pyproject.toml
[project]
name = "local-market-analyzer"
version = "0.1.0"
description = "Local-first AI market analysis terminal — bring your own LLM, run your own analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "yfinance>=0.2.40",
    "ta>=0.11.0",
    "langgraph>=0.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "sqlmodel>=0.0.38",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
